// LZEventApp/src/app/chat-app/graphql/operation-result-types.ts
export type getAllContributors = {
  getContributors:{
    _typename: "Contributors",
    email:string|null,
    paid:bool,
    notify:bool,
    amount: int,
  } | null,
}